refactored code to better repo structure. added a demo_dataset.ipynb in the examples folder. other demo files in progress. prediction codes are still in progress
