+++ campuscrowd/data_processed/SEQ.cfg
@@ -1,0 +1,14 @@
+[csv-path]
+flow_df_path = ./campuscrowd/data_processed/flow_df_SEQ_1fps.csv
+
+[CMGraph]
+adjacency_mat = 0,0,0,1,0,0,0,0,0,0
+                0,0,0,1,0,0,0,0,0,0
+                0,0,0,1,0,0,0,0,0,0
+                1,1,1,0,1,0,0,0,0,0
+                0,0,0,1,0,0,1,0,0,0
+                0,0,0,0,0,0,1,0,0,0
+                0,0,0,0,1,1,0,1,1,0
+                0,0,0,0,0,0,1,0,0,1
+                0,0,0,0,0,0,1,0,0,0
+                0,0,0,0,0,0,0,1,0,0
