+++ campuscrowd/data_processed/Stadium_2023.cfg
@@ -1,0 +1,11 @@
+[csv-path]
+flow_df_path = ./campuscrowd/data_processed/flow_df_stadium_2023_1fps.csv
+
+[CMGraph]
+adjacency_mat = 0,1,0,0,0,0
+                1,0,1,0,0,0
+                0,1,0,1,0,0
+                0,0,1,0,1,0
+                0,0,0,1,0,1
+                0,0,0,0,1,0
+
