+++ campuscrowd/evaluate_utils.py
@@ -1,0 +1,2 @@
+def evaluate():
+    return 0
