+++ campuscrowd/models/__init.py
@@ -1,0 +1,8 @@
+from .sten import DenseGCNGRU, GCNGRU
+from .temporal_only import GRU_only
+
+__all__ = [
+    'GRU_only', 
+    'GCNGRU', 
+    'DenseGCNGRU', 
+]
