+++ environment.yml
@@ -1,8 +1,7 @@
-name: pedestrian-test-2
+name: Campus-Crowd
 channels:
     - conda-forge
     - pytorch
-#     - nvidia
     - pyg
     - defaults
 dependencies: 
@@ -16,11 +15,8 @@ dependencies:
     - pytorch>=1.11
     - pyg
     - torchvision
-#     - pylibcugraphops>=23.02
     - pip
     - pip: 
-#         - pytorch-lightning
-#         - lightning-bolts
         - opencv-python
         - utm
         - filterpy
